feat: VIN-54 - Ajustes no AddressServiceImpl

+++ cadastro/src/main/java/com/vinhonotas/cadastro/application/services/impl/AddressServiceImpl.java
@@ -4,7 +4,11 @@ import com.vinhonotas.cadastro.application.converters.AddressConverter;
 import com.vinhonotas.cadastro.application.services.AddressService;
 import com.vinhonotas.cadastro.application.services.exceptions.BadRequestException;
 import com.vinhonotas.cadastro.domain.entities.AddressEntity;
+import com.vinhonotas.cadastro.domain.entities.CountryEntity;
+import com.vinhonotas.cadastro.domain.entities.StateEntity;
 import com.vinhonotas.cadastro.infrastructure.AddressRepository;
+import com.vinhonotas.cadastro.infrastructure.CountryRepository;
+import com.vinhonotas.cadastro.infrastructure.StateRepository;
 import com.vinhonotas.cadastro.interfaces.dtos.inputs.AddressInputDTO;
 import com.vinhonotas.cadastro.utils.MessagesConstants;
 import lombok.RequiredArgsConstructor;
@@ -12,6 +16,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -20,11 +26,27 @@ public class AddressServiceImpl implements AddressService {
 
     private final AddressRepository addressRepository;
     private final AddressConverter addressConverter;
+    private final StateRepository stateRepository;
+    private final CountryRepository countryRepository;
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public AddressEntity create(AddressInputDTO addressInputDTO) {
         try {
+            StateEntity state = stateRepository.findByUf(addressInputDTO.getUf().getUf());
+            if (Objects.nonNull(state)) {
+                addressInputDTO.setUf(state);
+            } else {
+                throw new BadRequestException(MessagesConstants.STATE_NOT_FOUND_WITH_NAME + addressInputDTO.getUf().getUf());
+            }
+
+            CountryEntity country = countryRepository.findByCountryName(addressInputDTO.getCountry().getCountryName());
+            if (Objects.nonNull(country)) {
+                addressInputDTO.setCountry(country);
+            } else {
+                throw new BadRequestException(MessagesConstants.COUNTRY_NOT_FOUND_WITH_NAME + addressInputDTO.getCountry().getCountryName());
+            }
+
             AddressEntity addressEntity = addressConverter.toEntity(addressInputDTO);
             return addressRepository.save(addressEntity);
         } catch (Exception e) {
@@ -34,7 +56,11 @@ public class AddressServiceImpl implements AddressService {
 
     @Override
     public List<AddressEntity> getAll() {
-        return addressRepository.findAll();
+        List<AddressEntity> addressList = addressRepository.findAll();
+        if (addressList.isEmpty()) {
+            throw new BadRequestException(MessagesConstants.ADDRESS_NOT_FOUND);
+        }
+        return addressList;
     }
 
     @Override
@@ -44,6 +70,7 @@ public class AddressServiceImpl implements AddressService {
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public AddressEntity update(UUID id, AddressInputDTO addressInputDTO) {
         try {
             AddressEntity addressEntity = this.getById(id);
@@ -54,7 +81,12 @@ public class AddressServiceImpl implements AddressService {
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public void delete(UUID id) {
+        Optional<AddressEntity> address = addressRepository.findById(id);
+        if (address.isEmpty()) {
+            throw new BadRequestException(MessagesConstants.ADDRESS_NOT_FOUND);
+        }
         try {
             addressRepository.deleteById(id);
         } catch (Exception e) {
